requirements.txt
streamlit
google-generativeai
requests
python-dotenv
langchain
langchain-community
faiss-cpu
transformers
scikit-learn